Split the maturin tests by concern after the rebase
Rebasing onto main combined two independent decompositions of the same
surface. Main extracted wheel inspection into
`tests/helpers/maturin_wheel.py` and added `maturin_script_locatable`
with three tests; this branch inlined the pin readers into their sole
consumer. Both landed, which left `test_maturin_build.py` at 446 lines —
past the 400-line limit, and mixing three concerns.

Split it the way the helpers already are: pin synchronization in
`test_maturin_pins.py`, the availability detectors in
`test_maturin_toolchain.py` beside the `toolchain_available` tests they
sit next to, and the wheel build and snapshot in `test_maturin_build.py`.

One reader resists inlining. `read_expected_maturin_version` now has two
consumers — the pin comparison and the wheel snapshot's `Generator`
assertion — which is exactly the second concrete consumer this branch's
own re-use policy names as the condition for sharing. It moves to
`_maturin_pin_support.py`; the readers with one consumer stay put.

Repoint `test_manylinux_container_ref_properties.py`, which imported the
container regex from the module that was split, and refresh the wheel
manifest for the new files.
